claim address widget in wallet
