+++ packages/commonwealth/client/scripts/views/pages/WalletPage/RewardsCard/RewardsCard.tsx
@@ -5,7 +5,6 @@ import { CWCard } from 'views/components/component_kit/cw_card';
 import { CWIcon } from 'views/components/component_kit/cw_icons/cw_icon';
 import { IconName } from 'views/components/component_kit/cw_icons/cw_icon_lookup';
 import { CWText } from 'views/components/component_kit/cw_text';
-import TokenClaimAddress from './TokenClaimAddress';
 
 import './RewardsCard.scss';
 
@@ -53,7 +52,6 @@ const RewardsCard = ({
         </CWText>
       )}
       <div className="rewards-card-body">{children}</div>
-      {title === 'Rewards' && <TokenClaimAddress />}
     </CWCard>
   );
 };
