--- packages/commonwealth/client/scripts/views/pages/WalletPage/RewardsCard/TokenClaimAddress.scss
@@ -1,24 +1,0 @@
-.TokenClaimAddress {
-  .claim-address-body {
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-
-    .claim-address-row {
-      display: flex;
-      justify-content: flex-start;
-      align-items: center;
-
-      .CWSelectList {
-        width: 100%;
-        .SelectList {
-          padding: 4px 8px !important;
-          line-height: initial;
-          min-height: initial;
-          max-height: 40px;
-          width: 100%;
-        }
-      }
-    }
-  }
-}
